Move PhysicsListBuilder and physics-list helpers into physics.py
